backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sse-starlette==1.8.2
yt-dlp>=2025.1.0
openai-whisper
notion-client==2.2.1
python-dotenv==1.0.0
httpx==0.26.0
